Move common utility functions to a separate file
